Json read codes fixed with extension, satellite positions operations added

+++ app/src/main/java/com/canerture/satellitesapp/data/model/Position.kt
@@ -1,6 +1,6 @@
 package com.canerture.satellitesapp.data.model
 
 data class Position(
-    val posX: Float,
-    val posY: Float
+    val posX: Double,
+    val posY: Double
 )
